Revision 1
Removed colored backgrounds and added fieldsets to show the sections for
each view we had planned. Also, put the data from the table into a text
file, the way Dr. Garfield had planned. The next thing I was working on
was trying to get the data out of the text file and into a table. It is
almost done, as I have saved each row of data into an array, and each
array into an array list. There is still much work to be done here for
retrieving the data, and I think much more thought and work needs to be
put into it.